pyproject.toml
[tool.poetry]
name = "npt-a"
version = "0.1.0"
description = ""
authors = ["devMuromi <devmuromi@gmail.com>"]
readme = "README.md"
packages = [{include = "npt_a"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.7"
django-rest-framework = "^0.1.0"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
